+++ packages/svelte/formula/src/lib/form/enrichment.ts
@@ -1,20 +1,17 @@
 import { FormulaOptions } from '../../types';
 
 /**
- * Create an enrichment object for the field
+ * Creates an enrichment object for the named group,
  * @param name
  * @param options
-
  */
-export function createEnrichField(
+export function createEnrichField<T extends {} = Record<string, unknown>>(
   name: string,
   options: FormulaOptions,
-): (value: unknown | unknown[]) => Record<string, unknown | unknown[]> {
-  return (value: unknown | unknown[]): Record<string, unknown | unknown[]> => {
-    const results: Record<string, unknown> = {};
-    for (const [key, fn] of Object.entries(options.enrich[name])) {
-      results[key] = fn(value);
-    }
-    return results;
-  };
+): (value: unknown) => T {
+  return (value: unknown): T =>
+    Object.entries(options.enrich[name]).reduce((a, [key, fn]) => {
+      a[key] = fn(value);
+      return a;
+    }, {}) as T;
 }