+++ packages/svelte/formula/src/lib/form/event.spec.ts
@@ -37,7 +37,7 @@ describe('Formula Event Handlers', () => {
 
       document.body.appendChild(el);
 
-      destroyHandler = createHandler('testing', 'change', el, [el], storeMock, {});
+      destroyHandler = createHandler('testing', 'change', el, [el], storeMock, {}, new Map());
     });
 
     afterEach(() => {
@@ -49,22 +49,28 @@ describe('Formula Event Handlers', () => {
       expect(destroyHandler).toBeInstanceOf(Function);
     });
 
-    it('should update the value when there is a change event', () => {
+    it('should update the value when there is a change event', (done) => {
       opts[1].selected = true;
       el.dispatchEvent(new Event('change'));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: 'B' });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: 'B' });
+          done();
+        })();
+      }, 0);
     });
 
-    it('should update the value when its a multiple change event', () => {
+    it('should update the value when its a multiple change event', (done) => {
       el.setAttribute('multiple', 'multiple');
       opts[0].selected = true;
       opts[1].selected = true;
       el.dispatchEvent(new Event('change'));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: ['A', 'B'] });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: ['A', 'B'] });
+          done();
+        })();
+      }, 0);
     });
   });
 
@@ -87,17 +93,20 @@ describe('Formula Event Handlers', () => {
       destroyHandler();
     });
 
-    it('should create a handler for a single checkbox', () => {
-      destroyHandler = createHandler('testing', 'change', el, [el], storeMock, {});
+    it('should create a handler for a single checkbox', (done) => {
+      destroyHandler = createHandler('testing', 'change', el, [el], storeMock, {}, new Map());
 
       el.click();
       el.dispatchEvent(new Event('change'));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: true });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: true });
+          done();
+        })();
+      }, 0);
     });
 
-    it('should create a handler for a multiple checkbox', () => {
+    it('should create a handler for a multiple checkbox', (done) => {
       const el2 = document.createElement('input');
       el2.id = 'test2';
       el2.type = 'checkbox';
@@ -106,13 +115,16 @@ describe('Formula Event Handlers', () => {
       el2.setAttribute('name', 'testing');
       document.body.appendChild(el2);
 
-      destroyHandler = createHandler('testing', 'change', el, [el, el2], storeMock, {});
+      destroyHandler = createHandler('testing', 'change', el, [el, el2], storeMock, {}, new Map());
 
       el.click();
       el.dispatchEvent(new Event('change'));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: ['test1'] });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: ['test1'] });
+          done();
+        })();
+      }, 0);
       document.body.removeChild(el2);
     });
   });
@@ -136,8 +148,8 @@ describe('Formula Event Handlers', () => {
 
       elements = [el1, el2];
 
-      const handler1 = createHandler('testing', 'change', el1, elements, storeMock, {});
-      const handler2 = createHandler('testing', 'change', el2, elements, storeMock, {});
+      const handler1 = createHandler('testing', 'change', el1, elements, storeMock, {}, new Map());
+      const handler2 = createHandler('testing', 'change', el2, elements, storeMock, {}, new Map());
       destroyHandlers = [handler1, handler2];
       document.body.appendChild(el1);
       document.body.appendChild(el2);
@@ -148,22 +160,28 @@ describe('Formula Event Handlers', () => {
       destroyHandlers.forEach((fn) => fn());
     });
 
-    it('should set the value when selecting a radio', () => {
+    it('should set the value when selecting a radio', (done) => {
       elements[0].click();
       elements[0].dispatchEvent(new Event('change'));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: 'A' });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: 'A' });
+          done();
+        })();
+      }, 0);
     });
 
-    it('should set the value when changing a radio', () => {
+    it('should set the value when changing a radio', (done) => {
       elements[0].click();
       elements[0].dispatchEvent(new Event('change'));
       elements[1].click();
       elements[1].dispatchEvent(new Event('change'));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: 'B' });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: 'B' });
+          done();
+        })();
+      }, 0);
     });
   });
 
@@ -184,33 +202,39 @@ describe('Formula Event Handlers', () => {
       destroyHandler();
     });
 
-    it('should create a handler for a single text boxes', () => {
-      destroyHandler = createHandler('testing', 'keyup', el, [el], storeMock, {});
+    it('should create a handler for a single text boxes', (done) => {
+      destroyHandler = createHandler('testing', 'keyup', el, [el], storeMock, {}, new Map());
 
       el.value = 'A';
       el.dispatchEvent(new KeyboardEvent('keyup', { key: 'A' }));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: 'A' });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: 'A' });
+          done();
+        })();
+      }, 0);
     });
 
-    it('should create a handler for a multiple text boxes', () => {
+    it('should create a handler for a multiple text boxes', (done) => {
       const el2 = document.createElement('input');
       el2.id = 'test2';
       el2.type = 'text';
       el2.setAttribute('name', 'testing');
       document.body.appendChild(el2);
 
-      destroyHandler = createHandler('testing', 'keyup', el, [el, el2], storeMock, {});
-      const secondHandler = createHandler('testing', 'keyup', el2, [el, el2], storeMock, {});
+      destroyHandler = createHandler('testing', 'keyup', el, [el, el2], storeMock, {}, new Map());
+      const secondHandler = createHandler('testing', 'keyup', el2, [el, el2], storeMock, {}, new Map());
 
       el.value = 'A';
       el.dispatchEvent(new KeyboardEvent('keyup', { key: 'A' }));
       el2.value = 'B';
       el2.dispatchEvent(new KeyboardEvent('keyup', { key: 'B' }));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: ['A', 'B'] });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: ['A', 'B'] });
+          done();
+        })();
+      }, 0);
       secondHandler();
       document.body.removeChild(el2);
     });
@@ -233,33 +257,39 @@ describe('Formula Event Handlers', () => {
       destroyHandler();
     });
 
-    it('should create a handler for a single number boxes', () => {
-      destroyHandler = createHandler('testing', 'keyup', el, [el], storeMock, {});
+    it('should create a handler for a single number boxes', (done) => {
+      destroyHandler = createHandler('testing', 'keyup', el, [el], storeMock, {}, new Map());
 
       el.value = '5';
       el.dispatchEvent(new KeyboardEvent('keyup', { key: '5' }));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: 5 });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: 5 });
+          done();
+        })();
+      }, 0);
     });
 
-    it('should create a handler for a multiple number boxes', () => {
+    it('should create a handler for a multiple number boxes', (done) => {
       const el2 = document.createElement('input');
       el2.id = 'test2';
       el2.type = 'number';
       el2.setAttribute('name', 'testing');
       document.body.appendChild(el2);
 
-      destroyHandler = createHandler('testing', 'keyup', el, [el, el2], storeMock, {});
-      const secondHandler = createHandler('testing', 'keyup', el2, [el, el2], storeMock, {});
+      destroyHandler = createHandler('testing', 'keyup', el, [el, el2], storeMock, {}, new Map());
+      const secondHandler = createHandler('testing', 'keyup', el2, [el, el2], storeMock, {}, new Map());
 
       el.value = '1';
       el.dispatchEvent(new KeyboardEvent('keyup', { key: '1' }));
       el2.value = '2';
       el2.dispatchEvent(new KeyboardEvent('keyup', { key: '2' }));
-      storeMock.formValues.subscribe((value) => {
-        expect(value).toStrictEqual({ testing: [1, 2] });
-      })();
+      setTimeout(() => {
+        storeMock.formValues.subscribe((value) => {
+          expect(value).toStrictEqual({ testing: [1, 2] });
+          done();
+        })();
+      }, 0);
       secondHandler();
       document.body.removeChild(el2);
     });