+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Types for Formula now provide stronger typing
+
+### Fixed
+
+- Formula now won't apply `aria-role` on items that already have a role set
+- Better handling of value and validation updates, store update is called less times
+
 ## [0.10.0] - 2021-03-12
 
 ### Added